strands-agents-examples/90-deployment/30-amazon-bedrock-agentcore/requirements.txt
strands-agents==1.13.0
strands-agents-tools==0.2.12

bedrock-agentcore==1.0.4
